Move sampleRowKeys over
